requirements.txt
requests 
bs4 
nltk 
fake-useragent
sentence-transformers
torch
pinecone-client
scipy